Move logger to utils package
