+++ dataset_loader.py
@@ -7,7 +7,7 @@ from dotenv import load_dotenv
 import torch
 from torch.utils.data import Dataset
 from transformers import BertTokenizer
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
 log.debug("Starting dataset_loader.py")
