+++ inference.py.orig
@@ -2,7 +2,7 @@ import torch
 from transformers import BertTokenizer
 from concept_inferencer import ConceptInferencer
 from vocab_manager import VocabManager
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
 log.debug("Starting inference.py.orig")
