+++ test/wopper_test.py
@@ -2,7 +2,7 @@
 # Basic smoke tests for the Wikidata interface and placeholders for future tests.
 
 from dotenv import load_dotenv
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
 log.debug("Starting wopper_test.py")
