+++ train_concept_model.py
@@ -1,7 +1,7 @@
 # train_concept_model.py
 # Trains the ConceptInferencer on the sample dataset with early stopping.
 
-from logger import get_logger
+from utils.logger import get_logger
 log = get_logger(__name__)
 
 log.info("Starting train_concept_model")
