+++ utils/__init__.py
@@ -1,0 +1,2 @@
+"""Utility modules for the wopper project."""
+
