+++ utils/hello_world.py
@@ -1,7 +1,7 @@
 # Minimal example module that prints 'hello world'.
 import os
 from dotenv import load_dotenv
-from logger import get_logger
+from utils.logger import get_logger
 
 log = get_logger(__name__)
 log.debug("Starting hello_world.py")
